+++ packages/form-builder/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @sio-group/form-types@0.4.1
+
 ## 0.3.1
 
 ### Patch Changes