+++ packages/form-validation/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @sio-group/form-types@0.4.1
+
 ## 0.1.6
 
 ### Patch Changes